[Patmos] Remove support for loopbound pseudo-instruction.
Old implementation used a pseudo-instruction, which was added
to loop headers with the loop's iteration bounds.
This instruction was then used by the Patmos target to generate
single-path code based on the loop bounds.

Removed the pseudo-instruction in preparation for using the
'llvm.loop' metadata node instead in the future.