[Patmos] TableGen: Remove reliance on custom operand skipping.
Previous implementation used "${p:skip" to allow the assembler
to skip the first operand in each instruction (the predicate).
This required custom support in llvm/utils/TabelGen/AsmMatcherEmitter.cpp

We remove this use from the Patmos TableGen and plan on implementing
the correct assembly parsing logic at a later point.